LayoutTests/media/track/track-webvtt-tc009-timings-hour-expected.txt

Tests cue timings that contain hours (they are optional), and tests various syntax errors in timings with hours.


*** Testing text track 0
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].id == '1') OK
EXPECTED (cues[0].startTime == '0') OK
EXPECTED (cues[0].endTime == '30.5') OK
EXPECTED (cues[0].text == 'Bear is Coming!!!!!') OK
EXPECTED (cues[1].id == '2') OK
EXPECTED (cues[1].startTime == '31') OK
EXPECTED (cues[1].endTime == '60.5') OK
EXPECTED (cues[1].text == 'I said Bear is coming!!!!') OK
EXPECTED (cues[2].id == '3') OK
EXPECTED (cues[2].startTime == '61') OK
EXPECTED (cues[2].endTime == '361200.5') OK
EXPECTED (cues[2].text == 'I said Bear is coming now!!!!') OK

*** Testing text track 1
EXPECTED (cues.length == '0') OK
END OF TEST
